refactor: add relative path to project
