+++ src/components/elements/simpleCard/style.js
