+++ src/components/pages/home/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { SimpleCard } from "../../elements";
+import { SimpleCard } from "@elements";
 
 const Home = () => {
   return <SimpleCard text="hello word"></SimpleCard>;
